Draft implementation
The Danske Spil API is called to retrieve a JSON object of all current
events. To make it easier to work with the quite complex JSON data
returned by the API, Rust types have been generated using typify by
handing it a JSON Schema generated by quicktype. This generation process
is incomplete and somewhat error-prone but a good starting point
nonetheless. Currently, the application simply prints the decimal odds
for the first event in the retrieved data.